Factor out story to snapshot story grid
